Support binding groups
There's a good chance that this isn't the implementation we want.

My brain is thinking that two values could be defined
far apart from each other, yet this would put them in the same location.

That would probably be pretty jarring.
But also, this is the first pass.
We can deal with that in the future.