Pretty print types
The pretty printer in the compiler has a different opinion about
what types should look like.
We're free to change that, but we have to implement it ourselves.
We go for a more compact format that should still be readable.

Postfixing the operators makes for cleaner diffs and
(subjectively) easier to understand type signatures.
Apparently Liam writes in a similar style.
And if Liam approves, it's gotta be good, right?