Push spacing out of declarations
The extra line was causing some weirdness.
If we have a declaration in an expression (like a `let`),
we don't want the extra line.

Sort of seems to make sense though.
A `Declaration` shouldn't always assume it's at the top level,
and so, should not always assume it needs an extra line.