Extract core functionality to library
This is in preparation for writing tests.
We could have just added the src directory to the tests,
but that usually ends up with extra rebuilds.
Rather, just split it along a decent seam.
Should also make it a bit clearer what we care about.

Changes `purty` around to return the stream for pretty printing.
We can print to a STDOUT or `Text` or whatever else we want.
In the tests, we probably want to print to `Text`.
We'll have to see what APIs are provided for us.