Diff with a program
If we don't choose the program to diff,
all we see in the output is a string representation of the module.
We don't get a diff at all!

We can choose the program to use for diffing.
And `tasty-golden` will output the diff of what's missing.
It's definitely no `hedgehog`, but it's better than nothing.