Start contribution guidelines
The process we use is a bit different if you're new to GitLab.
If this process ends up not working out well,
we should change it so it works better.

Also, maybe this file is improperly named.
We might want to rename it so it's clearer.
Or, maybe it's fine.