Outline the release process
Some of these steps and files don't exist yet.
This is mostly to get this stuff written down and out of my head.
They should either exist before merging this in completely,
or they should be removed from the process.

It's better to have `master` reflect the real state of the world
than be some utopia we'd like to get to.