Start running Travis CI
We want to build on OSX.
Travis is one of the few CI services that is free and has OSX builds.

The way we go about it is a bit convoluted:
* GitLab is our repo
* We mirror to GitHub
* GitHub send the build to Travis

It'd be nice if Travis would support GitLab out of the box.
For more information about GitLab and Travis,
see: https://github.com/travis-ci/travis-ci/issues/5931.