Use a newer version of `Win32-notify`
The `0.3.0.1` version has some compilation failure.
For more information about the problem,
see: https://github.com/haskell-fswatch/win32-notify/pull/1

Try to specify that we want a newer version if we need it.
This should only affect installs that actually need the dependency.
Which, for us ought to mean Windows builds only.