Use different files for bintray
The regexp didn't work as shown in the documentation.
Rather than spend time trying to figure out what's wrong with it,
we take the straight forward approach and copy the file.

This means we have one more round of versions to keep in sync.
We should really be generating these files, or not using this provider.