Run one stage in GitLab CI
Running CI in separate stages means spinning up a container each stage.
It's no biggie if the images are small, but `haskell:8.2.1` is large.

It takes a good two to three minutes to spin up, uncache and build cache.
If the image isn't there, it takes longer.
If the cache isn't there, it takes longer.
There's just a lot of overhead with using stages this way.

We can use stages, but maybe not for building the project.
Might be more useful for doing things separately with the built binary.