Send a notification on Travis CI start
It seems like we can remove the Zap from the GitLab side
if we get the notification at the start of the build.

Let's see how this plays out first before removing it completely.
It's possible the start notification is sent when the build is running.
That wouldn't really help us,
since we want the pipeline to wait until Travis is finished.
And if the pipeline finishes before Travis starts,
we've gained nothing.