Add a temporary file to the npm package
The ideal situation would be that `node-pre-gyp` would install `purty`.
The lifecycle for npm packages copies the binaries over before installing.
What this means, is that things fail when there is no `bin/purty`
to copy over.
We try to alleviate this situation by making this temporary file.
If someone tries to install, we'll first copy over this temporary file.
Then, `node-pre-gyp` should pull down our actual binary and replace it.
If `node-pre-gyp` can't pull it down,
this temporary file will fail when someone tries to use `purty` from npm.

This is kind of hacky, but meh.