Fix module path
We were pointing to the file.
But that meant we'd try to make a directory at `bin/purty`.
And you can't make a directory that's also a file name.
We'll, that's not entirely true, but it failed nonetheless.

What we do instead, is point at the directory `bin`.
This should mean we try to make a directory that already exists.
But since it should handle that case fine, everything should work.