Roll our own solution
The `node-pre-gyp` thing has too many arbitrary constraints.
Rather than continue massaging our project to work within it,
we just roll something simpler and clearer.

We pack the binaries in the package.
Every install has all platforms installed.
When someone calls `purty` from npm, we figure which binary to call.

This means the publishing process to npm has to be manual now,
but that's probably fine for a first go.