List which platforms we support explicitly
Should prevent someone from installing this on a platform we don't support.

This isn't a line in the sand we're unwilling to cross.
This is codifying the current state of the world.

We can try to add support for more platforms if someone requests it,
but this is fine to at least start with.