Fix dynamic type synonym newline test
We broke the static version in the process.
The static one was relying on the (weird) default layout options.
It was 80 characters with 40% ribbon width.
That was causing the formatting to be the way it was.

Moving to 80 characters and 100% ribbon width gives output you'd expect.
Rather than giving the weird 40% ribbon width to static,
we should fix the ribbon width to 100% and fix the static formatting.