Setup a unit test suite
We don't really need or want to have to go through the other test suites
for every assertion we want to make.
When we want to narrow our focus, we ought to be able to do so.

Using "small" assertions should help us do that.

This suite is failing on the static printing of empty record types.
That's expected, since they print differently.

What's unexpected is that it's passing for parenthesized empty records.
Probably we missed some case in the branching.