Print record puns
The compiler un-puns records are a mix between parsing and desugaring.
We have to take steps to re-sugar the AST.
We do that by taking control of printing binders.

We also print binders and exprs differently.
We don't want to indent the top level line.
We want to print out a single line with all the binders.

What we should probably do is make our own AST
that has some of this stuff figured out.
We might need to control the other printers we use from the compiler first.