Fix failing record pun tests
We have to "see through" positioned values.
Since basically everything is annotated with position and comments,
we have to look through all that and find the underlying value.

We _might_ want to handle parens,
but maybe not since we could remove necessary parens accidentally.
Again, we should convert to an AST that has all this stuff figured out.
It would make printing much easier.