Fix import sorting test
Just another example of why we should be using our own AST.
That we have to have partial matches hurts.
As soon as the AST changes in the compiler,
we have to hunt and peck to make sure we hit all of the cases.

In any case, this seems to work.
So, that's great!