Fix operator formatting
When we get to a `BinaryNoParens`,
we know that the operator...needs no parens.
So, we don't print that operator with parens.

Any other operator needs parens though, so we print them with parens.

What's telling in this commit is that we transform the structure
the same in both dynamic and static formatting.
What's different is how we handle the recursive case.