Extract `PurtyFilePath` from `Args`
The filepath sits in a weird place.
It's definitely needed from the user, as we can't give a good default.
But it's not really something we want in the configuration.

I think what this is showing is that `Args` is the wrong name here.
Probably, `Args` is the combination of the file path and the config.

Let's go with this for now.