Read config from a Dhall file
We want the Dhall file to be defaults.
If someone passes arguments on the command line,
assume they're more important than what's in the config file.

Otherwise, we fall back to what is in the config.

We also support a config file not existing.

Currently, we only support the config file being in the working directory.
We probably can look in a few places or take the location as an argument.

Either way, things to think about for the future.