Fix let bindings
It's nice that we could fix both problems in one location.
However, it's telling that we're still allowing *ALL* declarations in lets.
This includes syntax errors like a type class definition in a let binding.

We can pave over this here in our repo.
But really, the compiler shouldn't allow that sort of thing.