Separate arguments and configuration
There's a distinction between how the CLI is used,
and how the program is configured.

Up until now, they were effectively the same thing.
When we start adding more thing like spitting out the config,
and supporting STDIN, the distinction will be more pronounced.

We also made a semi-weird move with te filepath parsing.
We extracted it from `Args`, and smooshed it back together for parsing.
Really, that parsing should be part of `Args`.