Print out default configuration
This should provide a nice starting point for people unfamiliar with
either Dhall or `purty`.

It might be nice to provide some comments in the Dhall file.
Unfortunately, I think that means we'd either have to:
write dhall directly, or create an `Expr Src X` with comments.

Neither sounds appealing.
The former will get out of sync.
The latter will be hard to write.