Convert to classy prisms
Dual to the classy lenses we use, we start using classy prisms.
The idea being that we want to allow extensibility in the errors we have.

There's a good chance this won't turn out well.
But, it should make it a bit easier to work with the different
errors we'll soon have.