Start converting to our own AST
We're finally starting the process of writing an AST closer to our problem.
The PureScript AST worked well to get us off the ground,
but it has some quirks that made using it less than ideal.

It flattens out many levels of the syntax into fewer.
For instance, a let binding can have any declaration as a binding.
Which means that a let binding could contain an instance declaration.
That's no valid syntax in the slightest, yet we had to deal with it.

It has other quirks, but we've talked about them in depth in other areas.
What's important in this commit is that we are moving to an ASt
that is more correct for our purposes.
If the AST in the compiler works for the compiler, great!
But, it wasn't working out as well for us.

So far, we only have the module name.
Even that has made things more sensible by converting to a module name
that always has at least one name.
In the PureScript AST,
it's possible to have a module name without any names.

We also take this time to fix the missing newline for modules.
Hopefully, this won't blow up in our face...