Start implementing exports
We're taking a slightly different approach from before.
We're making the multiple vs. single line difference more explicit.
This change should help when we get to doing interactive formatting.

We also try something a little different with `NotImplemented`.
Makes it a bit more explicit what we've done and not done.
Though, not that much more explicit...