Support types exports
The weird part here is that PureScript represents everything as `Nothing`
and nothing as `Just []`.
It might make sense in the context of PureScript,
but it's really confusing outside of that context.

Also, the sorting annotation doesn't really work.
We have to remember to do it rather than it being caught statically.
Again, the problem is that sorting isn't algebraic.
At least, not the way we've defined it.