Handle other exports
These other two shouldn't be possible.
They're not possible to write in the source language,
but things that the compiler inserts at some point.

This mixing of concerns is a prime example of what we want to get away from.
The compiler is free to mix concerns all it wants,
but it affects us if we try to use the library at face value.
Making this layer that removes the mixing is the next best thing to
writing a parser and keeping it in sync with the compiler.

In other news, the prisms are working out really well.
Between the granularity of each prism and `-Wredundant-constraints`
we get alerted when we say we can return an error we actually can't.
In this case, we caught when `IsNotImplemented` stopped being returned.
Prisms!!!