Unify most of the formatting
This will end up being a breaking change to the output.
But, it's better than trying to maintain two nearly identical formats.
It should also make things a bit easier come the interactive format.

For the sake of effort, we could probably just move to interactive now.
But for the sake of sanity,
it's probably best to wait until the AST is done.