Move make exports its own type variable
We can speak more precisely about exports if we allow it to be different.
The annotation isn't really working the way it was originally intended.
Partly because sorting isn't an easy thing to express.

We end up having to write quite a bit to switch data structures,
but that's not a bad thing.
It keeps us more honest than before.