Handle all import declarations explicitly
We can gain more confidence changing PS versions if we're more explicit.

If something is added or removed, we'll know about it.
If something we do care about (import) is changed,
we keep the original behavior.
If something we don't care about (non-import) is changed,
we won't get noise as we're using record syntax.