Start converting data and newtypes
There are lots of intricacies about the PureScript types
which makes this less than straight forward.
We're still too close to the type form the PureScript library.
We should make a couple passes to move away from that.

Given how long this took, probably not anytime soon unless it's urgent.
There's quite a bit more of the language to handle.