Format `if` expressions differently
Do notation changes the the parse rules for what's valid.
If we print an `if` expression as:

```
foo = do
  if true then
    1
  else
    2
```

The parser will see the `if` and `else` keywords on the same column,
and think that `else` is starting a new bind statement.
Since it hasn't finished parsing the `if` expression,
it will fail to parse this section.
We change the way `if` expressions print to address this issue.

Now the above formatting would be:

```
foo = do
  if true
    then 1
    else 2
```

It might not be ideal for some branches, but at least it shouldn't break.