Switch over to extensible effects
The money here is how much we don't have to do with classy optics.
For the lenses, we don't have to have as many data types.
We also don't need any type classes.

The change isn't that drastic for the prisms,
since we only had two errors and not a full hierarchy yet.
Once we have more errors it'll cut down on a ton of boilerplate.

Overall, it seems pretty nice and has cut down quite a bit on boilerplate.