Move parsing to its own module
This is going to conflict with the changes in the AST branch,
but it makes sense currently.
The parsing takes a bit of lining up that isn't core to the problem.
We can dump it into `Module` for now, and not worry too much about it.

Maybe parsing deserves its own module?
It seems unlikely since we only use the PS parser for modules.