Support fixities
This is fleshing out some inconsistencies with declarations.
Here starts the weird stuff.
Hopefuly, we can find a nicer way to format each declaration.
It feels like we have to give and take to get the right thing formatted.

Also, the annotation for `Name.TypeOperator a` and `Name.ValueOperator a`
didn't work out so well.
Seems like this is more that the annotations aren't right
more than anything else.

For one thing,
the only time braces makes sense is in the context of records.
But `Annotation.Normalized` can be used everywhere.
Probably, we should be more granular with the annotations.
Also, `Annotation.Sorted` doesn't really make much sense.

Might take a while to get where we need to be.