Fix rows
While the cons list representation might be elegant,
it makes it much harder to format things properly.
We didn't know when something should have a comma after it,
when a row was open until it was too late, etc.

We convert to a more abstract list representation so easier to work with.

We also format them so they're valid syntax.