Support let bindings and where clauses
We end up supporting quite a bit more as well to make the tests pass.
Seems like this is one of the places where the dynamic/static
is really prevalant.

We're not entirely honest about the expressions.
For instances, we say that in a `DoExpression` it can be any `Expression a`,
but that's not really true.
It can't be a where clause.

Really, where clauses aren't expressions.
But PS lumps them in as expressions.

We should split them off to being the syntax they are.