Handle parens in types better
We were letting everything format how it wanted to even in parens.
It doesn't work out too well for things with functions within them.

Instead, we assume that if parens are there,
the thing should be formatted on a single line.

This is more fuel for supporting an interactive format,
as we wouldn't have to try and handle this case.