Generate the Bintray JSON files from Dhall
We can skirt around two of the manual steps when making a release.
We used to have to change the versions and date in both of these files.
While it's not hard work, it's tedious and error prone.
We have computers for this sort of work.

We also make a change to always publish whatever binaries we upload.
So far, we haven't had an issue where the binary was broken.
It seems a safe bet to default to publishing,
and handle problems as they arise (instead of the other way).