Ensure label check is on normalized values
We were checking un-normalized values and missing out on most puns.
The PS parser puts a bunch of comment annotations on things,
even if there are no comments.
And we weren't able to see through the comments.
Now, we normalize away empty comment things, so we don't deal with it.

We also force the check to happen on a normalized thing.
While we don't really care about the annotation,
we do want to make sure we're looking at as far normalized a thing.