Prefer `Member` over `Members`
We had a couple of redundant constraints that were hidden by `Members`!

When we use `Members`,
GHC cannot perform analysis of redundant constraints.
The underlying constraint `Member` is a plain old constraint;
if we use it, GHC can do redundant constraint analysis.

We should prefer using `Member` as much as we can, for this reason.