Test against an "original" file
In an effort to show more clearly how the formatting works,
we keep the "original" file around.
This is supposed to be the kind of file a normal human would write.

It should give someone an easier way to see what turns into what.
Will work really nicely if we cover a good amount of the language.
Will also be really nice once we get to interactive formatting.