Make the "original" files the source of truth
To try and ensure we support formatting for both versions,
we start from the "original" file and check it against the output.

If we think of the golden test as mapping `original -> dynamic/static`,
this suite was surjective before, but now is now injective.
Ideally, the suite would be bijective.
Otherwise, we'll end up with dead files.