Start over with interactive formatting
This past week has been an adventure.
Over the last week, the entire codebase was re-written.
Things went from ~5k lines of Haskell to ~3k lines.
Formatting was using the interactive style.
Life was good.

As things were being cleaned up today to make a commit,
the command `find . -delete -name '*.tix' -type f` was run.
That command doesn't delete all of the `*.tix` files recursively.
That command deletes all of the files, period.

All of the work of the past week was gone.
Tried a few things to recover the files, but no dice.

Rather than be angry, we move on and start anew.
This time though, we're pushing to the remote repo often.

Things might go quicker this time since this is re-write #2 in
as many weeks.
Or, it'll just take the same amount of time because it's a large project.

In any case, the changes here take a simpler approach than previous.
We embrace the fact that `IO _` will have to be used at some point,
and set us up with a framework that uses `IO _` pervasively.
The trick with this method is to make dependencies explicit
and only use them as the way to interact with `IO _`.

We could probably make it `f _` and pass around constraints and whatever.
But, we end up with easier to understand types and errors if we just assume
that `IO _` will be used responsibly.
Given that the number of people actively developing in the repo is ~1,
that seems like a fair trade-off to make.