Handle exitcodes properly
`ComponentM _` will fail if we use `exitSuccess`,
since `exitSuccess` throws an exception in order to exit.
We have to leave the callback gracefully,
and exit when we're back at the top level.

This is actually a really good thing.
It pushes us to have exactly one exit point from our program.
That should make the behavior of the program easier to understand.