Flesh out errors a bit more
We would like to follow an approach to errors where we make exactly
one decision about them.
When we get an error, we've got two choices: handle the error or
let the caller deal with it.

We don't log any errors when they happen as that would be making a
decision about how to handle the error.
Instead we propagate any errors we cannot handle to our caller and let
them deal with handling it.

If an error reaches the top-level, it means that it's fatal.
We communicate the error and exit with a failure code.