Scaffold out more of the structure
One lesson learned from the implementation last week is that there are
only a handful of modules we need to make a functioning program.
Splitting across behaviors (rather than types) seemed to make for a
more cohesive set of fewer modules.
We try to follow that approach here as well.
There were three real big things we found last week:
* We need to know the span of a construct.
* We need to know how to annotated a construct.
* We need to know how to format a construct.

While annotation should be really straight-forward,
it's actually quite a bit more tedious than it should be.
We can make it less-so by leveraging span knowledge.

Formatting is relatively straight-forward, which is surprising.
There are a couple of quirks, but overall it's a fold over the tree.