Start formatting modules
We only tackle the module header at the moment.
We have to flesh out a couple of the constructs we'll use later,
but most of the constructs here pay-off the least of the whole module.
Most of what's happening here is one-off syntax for module headers.
Just one of those things we have to do.