Format imports
We don't bother to sort things or separate them by type at the moment.
There's a lot of thrashing that `purs ide` will do with us if we do.

It would be nice if we had the same algorithm for imports,
but it seems like our opinions differ on how best to format imports.