Fix formatting of cases
We change the `OperatorPattern` formatting as a shortcut.

Ideally, we'd always have a newline for cases.
The intent there is that we minimize diffs if we don't thrash between
single line cases and multiple line cases.

Unfortunately, the formatting algorithm isn't great and handling cases
differently is harder than it seems.