Format `LetBinding _`s correctly
We had to pass around the prefix and suffix.
This seems more complex than the previous implementation.
It works though, so that's all that really matters.

We also get back single line let bindings.
It's unclear if this is a good thing or a bad thing.

This also marks the point where all the golden tests pass again!
We can start making changes in smaller chunks and whatever.