Add back support for configuration
We continue to have the same motivation behind configuration.
It's not to open the flood gates to configuring formatting.
It's to make a common workflow easier to deal with.
In particular, if you always want to write to files in-place,
you shouldn't have to always pass a `--write` flag.

That motivation might change in the future.
If it does, most likely it will be to remove configuration all together.