Always convert a filename to an absolute filename
Looks like when we use `purty` from a different directory
(could be found on the `$PATH`) it will not find the file appropriately.
It thinks the file is relative to the `purty` binary.

Instead, we attempt to make the filename given into an absolute filename.
This should allow `purty` to work no matter where it's located.