Move argument parsing into `Args` module
We don't have to expose `Main` to the internals of `Args`.
In particular,
it shouldn't matter that we're using `optparse-applicative` in `Args`.
That is 100% an implementation detail.

What matters is that `Args` can create values that `Purty` can consume.
We abstract over how to parse args, and push it into `Args`.