Move more logic into `Purty`
We don't need to talk about the intricacies of how to run the program.
So long as we can give the runner the arguments it needs,
it ought to be able to do all the running and work normally.

We're able to drop more dependencies from the executable,
and focus more on what it should be doing: running the program.
It might even still be doing too much,
but this is probably a decent place to stop.