Handle more failures in the tests
`runComponentM` catches exceptions and does its own handling.
Looks like if `goldenTests` were to fail, it would not fail the test suite.
But, we want this to fail loudly if anything goes sideways.
We make the whole block work like normal, convert any exceptions to values,
and process them outside of the `ComponentM _` block.

This should make for more reliable test failures.