Fix the formatting of `ExprApp _ _ _`
We weren't formatting correctly when an `ExprApp _ _ _` spanned
multiple lines, but the second `Expr _` spanned a single line.
The issue was that we were determining what to do in `Format.exprPrefix`
based on the second `Expr _` we were given.
In reality, we need to look at the entire `ExprApp _ _ _` span,
and determine how to format.