Add failing tests for spacing in arrays and records
We want to be a bit more consistent with the spacing for arrays and records.
In particular, we want the spacing to be similar between single and
multiple line versions for the formatting.
to facilitate that, we add extra spaces on the single line versions.

The changes here should be indicative of what needs to be fixed before
this issue is considered complete.