Format conditionals the old way
We used to format conditionals where the `then` and `else` branches
would be indented the same amount:
```PureScript
if foo
  then bar
  else baz
```

Now that we're using on `0.13.x` for `purescript`,
we can go back to the format we used before:
```PureScript
if foo then
  bar
else
  baz
```

The latter is more verbose,
but should be more in-line with the goals of this project.
E.g. if `bar` needs to change, it shouldn't affect the line above it now,
whereas before it might do that depending on what `bar` was.