Handle missing files
If a file doesn't exist, we don't need to explode when trying to read it.
Missing files is not an exceptional situation.
That's a normal occurence of working with an evolving file system.
People typo file paths, files move, etc.

Instead of treating this as an unrecoverable situation,
we try to read the file, and if some `IOException` happens, we handle it.
It would be nice if there were a `withBinaryFile` equivalent that
did this logic itself (so we didn't have to catch _every_ `IOException`).
This seems like the best we can do with what we've got though.