Support reading from STDIN
Something about this feels off.
We don't really want to allow the `--write` flag if we're using STDIN.
What are we writing to?

If we try to separate the `Format` into different cases,
it looks impossible to parse `-` as STDIN while also parsing any string
as a file path.

Probably these should be subcommands (in `optparse-applicative` jargon).
That change would be breaking for the CLI,
and it doesn't seem worth it right now.
Breaking changes aren't a problem,
but this was supposed to be a small extension.

We can break it later when we're feeling more frisky.