State name explicitly when reporting an error
We handle STDIN differently than any old file path.
In both cases, we communicate what we were working with.

Since we have exactly one exit point,
we can handle this logic in exactly one place.
This is the nice thing about not having multiple exit points;
we don't have to litter the code with logic to handle the same case.