Choose a specific `get-stack.sh` to download
Earlier this week, `stack` released a new version.
That version changed the file, our CI process recognized it and failed.
Since we are not quite ready to update to 2.1.1
(there's quite a few broken things being reported for that version),
we choose an older version and use that.

Unfortunately, we didn't reference the actual file before.
So we have to choose one of these arbitrarily.
We checked all the commits from version 1.9.3,
but none of the SHAs lined up with the SHA we had here.
Something wonky is going on there.

In any case, this is the latest 1.9.3 version.
So, that should hopefully work alright.