Add missed changes for constraints
Had a slip while making the previous commit.
We forgot to commit this part.
The message is the same as the previous commit.

It does sort of highlight that we're doing shotgun surgery here.
It's a code smell for a reason.
We should consider handling parenthesized constructs uniformly,
so this sort of mistake is mitigated.