Handle parenthesized constructs uniformly
To address parenthesized constructs getting out of sync with each other,
we extract a function that has that behavior.
This should keep things more uniform
and mitigate the amount of shotgun surgery we have to perform.

Things aren't typed to the point that changes to any parenthesized
constructs would be caught statically,
but hopefully this is a good enough compromise.