Add a failing test for typed binders
Looks like we weren't formatting out typed binders appropriately.
We are not putting a space between the binder and the colons.

Interestingly, we seem to be formatting typed expressions correctly.
It sees like we could slice the code in a way that these two things
were formatted the same way.
Reason being, they're semantically very similar.
The fact that they're different types isn't worth much at the user-level.