Fix formatting of characters
There was a bug in the `0.13.0` version of `purescript`.
Control characters were not being parsed correctly.
To get around that bug, we decided to use the underlying `Text`
instead of relying on the parser's source value.

Unfortunately, that meant that we created a new bug.
Because the underlying `Text` value was converted to the escape code,
Certain characters (like unicode) were formatted as escape codes
rather than the original characters that they were.

Worse still, it surfaced an additional bug in the `0.13.0` version
of `purescript`: other escape code characters were not escaped properly.

The long and short was that there wasn't an easy way to fix the issue.
We decided to stave off until a new version was released.
Now that we have `0.13.2` (the next version available),
we can fix the original issue once and for all.