Set `LC_ALL` to work around `language-javascript`
`language-javascript` decided to add a single unicode character.
That one character breaks anything downstream of it if the locale
is not set appropriately:
https://github.com/erikd/language-javascript/issues/86.

The resolution in the above issue is that the one character will not
be changed and that compiling `language-javascript` requires a
UTF-8 compatible encoding.
Since it looks like we've no other option if we want to use this version
but to change our locale, we do so.