Fix indentation of conditionals
We were formatting conditionals less than ideally.
We would format conditionals where they weren't always indented properly.
For example, we would format conditionals like:
```PureScript
foo = do
  x <- if true then
    pure 1
  else
    pure 2
  pure x
```

While that will technically parse,
and it has the semantics it's supposed to,
It's not readily apparent that it should do either of those things.

Instead of expecting people to understand the parser (and what it expects),
we change the formatting to be:
```PureScript
foo = do
  x <-
    if true then
      pure 1
    else
      pure 2
  pure x
```

This formatting is a bit easier to understand.
The conditional is indented as it normally is in other places.