Fix indentation for multiple line instance heads
We were formatting all instance heads with a single space between types.
That works fine for single line instance heads.
It's not ideal for multiple line instance heads.

Before, we were formatting like:
```PureScript
instance fooABPositive ::
  Foo
    A ( B
      Positive
  ) where
```

After the changes in this commit, we are formatting like:
```PureScript
instance fooABPositive ::
  Foo
    A
    ( B
        Positive
    ) where
```

Although the example might look verbose, it shows the structure better.
The formatting is more consistent with how it is in other places.