Update copy in the README
We've done quite a bit of work in this project.
`purty` is no longer "work in progress."
It's a useful program and the way it formats is nice.
We should be a bit better about how we speak of the work we've done.

We also update usage and language around what `purty` does.
It's a source code formatter.
It's not a "pretty printer."

Finally, we update the link to the Dhall tutorial.