Add golden tests to the `test` target
We make the golden tests a separate target,
but also make it a dependency of `test`.
This way, if someone calls `make test`, they will run all the tests.
If they only want the golden tests, they can use `make test-golden`.
We use the `test-golden` target for Appveyor instead of the `test` target
as it only runs the golden tests.