Go back to building the command as a variable
It's arbitrary whether we use flags in the recipe or in a variable.
We go back to recipies being easier to understand.

It seems like people have already run into this issue in the past
and come to some conclusions about how best to deal with flags.
We should find some of those conclusions and base our decision on them.
No point going through pain if best practices already exist.