Flesh out the rest of the test targets
Many changes in this commit.
First, we add a rule to build the binaries.
We want a rule to build the binary as is.
Next, we want a rule to put the binary in a platform-specific location.
Finally, we make the dependencies of the `test*` targets more accurate.

These changes should make for a more accurate build when things change.