Update `purescript` to 0.13.6
We're a few versions behind. We bring up to the latest version so we can
try to stay current. We might want to think about figuring some way to
stay up to date better. The hard part is that `purescript` has no set
release schedule, so we'd have to pull more frequently than necessary in
order to be up to date.