Download `dhall-to-json` instead of using `stack`
Up until now, we've used `stack` to manage our binaries. This was
motivated by two things: `dhall-json` didn't have a pre-build macOS
binary, and `stack` makes it pretty straight forward to install Haskell
binaries.

The downsides to using `stack` are that you have to keep dependencies in
check for something that should be tooling. In our case, `dhall-json`
depends on `dhall` and on a specific version of it. Since we don't want
to depend on `dhall` anymore, we move over to downloading the pre-built
binaries instead of building them from source.

We might flip back on this decision if things change. For now, it seems
like a good way forward.